ela_rag_docker/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dash==3.0.4
et_xmlfile==2.0.0
faiss-cpu==1.10.0
fastapi==0.115.11
filelock==3.17.0
Flask==3.0.3
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.43
langchain-text-splitters==0.3.6
langsmith==0.3.13
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
plotly==6.0.1
plotly-express==0.4.1
proto-plus==1.26.1
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.5
pyparsing==3.2.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
psycopg2==2.9.10
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.1
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.0.6
zipp==3.21.0
zstandard==0.23.0